+++ src/cli/node/state/index.js
@@ -1,6 +1,7 @@
 const shell = require('shelljs');
 const chalk = require('chalk');
 const { cli, action } = require('core/cli');
+const { getTopRunningChains } = require('core/libs/common');
 const { execute, run } = require('./state');
 
 cli('status [type]', 'List info of the running chain/node', input => action(execute, run, input), {
@@ -9,6 +10,8 @@ cli('status [type]', 'List info of the running chain/node', input => action(exec
     runningNode: true,
     rpcClient: true,
     wallet: false,
+    currentChainRunning: true,
+    chainName: getTopRunningChains,
   },
   options: [],
   handlers: {
@@ -21,6 +24,7 @@ Examples:
   - ${chalk.cyan('forge status net')}       display status of network
   - ${chalk.cyan('forge status validator')} display status of validators
   - ${chalk.cyan('forge status all')}       display status of all components
+  - ${chalk.cyan('forge status -c hello')}  display status for chain named hello
         `);
     },
   },